# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "College API Documentation",
        "description": "API sučelje za upravljanje studentima, odjelima i profesorima u sustavu fakulteta.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Lokalni razvojni server"
        }
    ],
    "paths": {
...
        "/api/faculty": {
            "get": {
                "tags": [
                    "Profesori"
                ],
                "summary": "Dohvati listu profesora",
                "description": "Vraća paginiranu listu svog nastavnog osoblja.",
                "operationId": "getFacultyList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Broj stranice za paginaciju",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspješno dohvaćanje profesora",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "first_name": {
                                                        "type": "string",
                                                        "example": "Pero"
                                                    },
                                                    "last_name": {
                                                        "type": "string",
                                                        "example": "Perić"
                                                    },
                                                    "department_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "061123456",
                                                        "nullable": true
                                                    },
                                                    "department": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Računarstvo"
                                                            }
                                                        },
                                                        "type": "object",
                                                        "nullable": true
                                                    },
                                                    "is_hod_of_department": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Računarstvo"
                                                            }
                                                        },
                                                        "type": "object",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Neautoriziran pristup / Potrebna autentifikacija.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Profesori"
                ],
                "summary": "Kreiraj novog profesora",
                "operationId": "storeFaculty",
                "requestBody": {
                    "description": "Podaci za novog profesora",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "department_id"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "maxLength": 25,
                                        "example": "Novi"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "maxLength": 25,
                                        "example": "Profesor"
                                    },
                                    "department_id": {
                                        "description": "ID postojećeg odjela",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "phone": {
                                        "type": "string",
                                        "maxLength": 10,
                                        "example": "062333444",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profesor uspješno kreiran",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "first_name": {
                                            "type": "string",
                                            "example": "Novi"
                                        },
                                        "last_name": {
                                            "type": "string",
                                            "example": "Profesor"
                                        },
                                        "department_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "062333444",
                                            "nullable": true
                                        },
                                        "department": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Računarstvo"
                                                }
                                            },
                                            "type": "object",
                                            "nullable": true
                                        },
                                        "is_hod_of_department": {
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Greška validacije",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "first_name": [
                                                    "The first name field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Neautoriziran pristup / Potrebna autentifikacija.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "basicAuth": {
                "type": "http",
                "scheme": "basic",
                "description": "HTTP Basic Authentication. Koristite email kao korisničko ime i odgovarajuću lozinku."
            }
        }
    },
...
}